🐛 Only up the server once database is connected
- Await until the connection with the database is etablished before up the server
- Move the server logic into a dedicated server/ folder
- Create an index file at the root of src/ that loads the server configuration and up it
